Updated user roles and enums in api_models and diesel_models
